requirements.txt
flask
sqlalchemy
werkzeug
mysqlclient
simplejson  # Fixes Decimal (de)serialization
python-dateutil
requests

click
